third_party/blink/web_tests/http/tests/devtools/tracing/tracing-timeline-load-expected.txt

Tests tracing based Timeline save/load functionality.


Running: testNormal
TimelineLoaderClient.loadingStarted()
TimelineLoaderClient.processingStarted()
TimelineLoaderClient.loadingComplete(true)
Saved data is equal to restored data: true

Running: testJSONObjectFormat
TimelineLoaderClient.loadingStarted()
TimelineLoaderClient.processingStarted()
TimelineLoaderClient.loadingComplete(true)
Saved data is equal to restored data: true

Running: testJSONObjectFormatWithMetadata
TimelineLoaderClient.loadingStarted()
TimelineLoaderClient.processingStarted()
TimelineLoaderClient.loadingComplete(true)
Saved data is equal to restored data: true

Running: testBroken
TimelineLoaderClient.loadingStarted()
TimelineLoaderClient.processingStarted()
TimelineLoaderClient.loadingComplete(true)
Model is empty: true

Running: testMalformedJSON
TimelineLoaderClient.loadingStarted()
TimelineLoaderClient.loadingComplete(false)
Model is empty: true
